Add files via upload
"I am a sergeant on a military peacekeeping mission. My task is to drop a package in a designated red drop zone. 
The package contains mission-critical items and is very important to be delivered at the exact location for the success of the mission.
Everyone before me has failed to finish the task. The success of the mission now depends on me."
No pressure. ;)